Working on UI TEmplate page #10
Refactor Mailings-Templates for improved UI/UX

Replaced static grid-based layout with a dynamic design using Syncfusion components (`SfDropDownList`, `SfComboBox`, `SfRichTextEditor`). Updated `Mailings-Templates.razor` to include a `groupbox` layout for better organization and added metadata fields for templates.

Refactored `Mailings-Templates.razor.cs` to handle template and recipient list selection, added a rich text editor toolbar, and implemented resource management with `CancellationTokenSource`. Deprecated unused `SfGrid` methods.

Updated `Mailings-Templates.razor.css` with styles for the new layout and editor. Modified `RecipientListModel` to support nullable `ListId` for better flexibility. Minor type compatibility fix in `Mailings-Recipients.razor.cs`.

+++ Amusing/Components/Pages/Mailings-Recipients.razor.cs
@@ -636,7 +636,7 @@ public class MailingsRecipientsBase : ComponentBase
             return;
         }
 
-        await MailingService.DeleteRecipientQueryAsync( SelectedRecipientsList.ListId );
+        await MailingService.DeleteRecipientQueryAsync( ( uint )  SelectedRecipientsList.ListId  );
 
         // Refresh the list
         RecipientsList = await MailingService.GetRecipientListsAsync();
